+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
-## [[unpublished]](https://github.com/mlange-42/arche/compare/v0.1.1...main)
+## [[unpublished]](https://github.com/mlange-42/arche/compare/v0.1.2...main)
+
+Nothing
+
+## [[v0.1.2]](https://github.com/mlange-42/arche/compare/v0.1.1...v0.1.2)
 
 ### Other
 
-* Use aligned item size in component storage, for approx. 20% faster query iteration (#25)
+* Use aligned item size in component storage for faster query iteration (#25)
 * Queries lock the World, and automatically unlock it after iteration (#26)
 
 ## [[v0.1.1]](https://github.com/mlange-42/arche/compare/v0.1.0...v0.1.1)